XWayland: move SocketFd functionality into XWayland spawner class/file
