+++ src/server/frontend_xwayland/xwayland_spawner.h
@@ -1,0 +1,56 @@
+/*
+ * Copyright (C) 2018 Marius Gripsgard <marius@ubports.com>
+ * Copyright (C) 2019-2020 Canonical Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 or 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#ifndef MIR_FRONTEND_XWAYLAND_SPAWNER_H
+#define MIR_FRONTEND_XWAYLAND_SPAWNER_H
+
+#include <memory>
+#include <string>
+#include <vector>
+#include <functional>
+
+namespace mir
+{
+class Fd;
+namespace frontend
+{
+
+class XWaylandSpawner
+{
+public:
+    XWaylandSpawner();
+    ~XWaylandSpawner();
+
+    XWaylandSpawner(XWaylandSpawner const&) = delete;
+    XWaylandSpawner& operator=(XWaylandSpawner const&) = delete;
+
+    auto x11_display() const -> std::string;
+    auto socket_fds() const -> std::vector<Fd> const&;
+
+    /// Returns if the operation succeeded
+    static bool set_cloexec(int fd, bool cloexec);
+
+private:
+    int const xdisplay;
+    std::vector<Fd> const fds;
+};
+
+}
+}
+
+#endif // MIR_FRONTEND_XWAYLAND_SPAWNER_H
